修改首页分类变量类型和xml文件

+++ src/main/java/medicine/online/client/backend/controller/CategoryController.java
@@ -29,10 +29,9 @@ public class CategoryController {
 
     @GetMapping("/categorylist")
     @Operation(summary = "首页分类子列表")
-    public List<CategoryVO> getCategoryListByParentId(@RequestParam(required = false) Long parentId) {
+    public List<CategoryVO> getCategoryListByParentId(@RequestParam(required = false) Integer parentId) {
         return categoryService.getCategoryListByParentId(parentId);
     }
-
 
 }
