+++ src/main/java/medicine/online/client/backend/mapper/CategoryMapper.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface CategoryMapper extends BaseMapper<Category> {
 
     // 自定义方法，根据parentId查询Category列表
-    List<Category> selectCategoriesByParentId(Long parentId);
+    List<Category> selectCategoriesByParentId(Integer parentId);
 }
